fix: review findings — BigDecimal tolerance, rollback null-guard, GetAgentAuditLogUseCase
- ReconcileEntriesCommand.tolerancePercent: Double? → BigDecimal? to avoid
  IEEE 754 imprecision in financial tolerance arithmetic; MCP boundary converts
  Double? with BigDecimal(it.toString()) for exact representation
- RollbackWorkflowService: replace silent return@forEach on null transactionId /
  missing original tx with checkNotNull/error — partial rollback now fails fast
  instead of returning ROLLED_BACK with silently skipped steps
- Extract GetAgentAuditLogUseCase (query + interface in application/agentic,
  GetAgentAuditLogService in infrastructure) so IdemMcpServer no longer injects
  AgentAuditRepository directly; consistent with all other MCP tool delegations

+++ application/src/main/kotlin/finance/idem/application/reconciliation/ReconcileEntriesCommand.kt
@@ -2,6 +2,7 @@ package finance.idem.application.reconciliation
 
 import finance.idem.core.AccountId
 import finance.idem.core.TenantId
+import java.math.BigDecimal
 import java.time.Instant
 
 data class ReconcileEntriesCommand(
@@ -9,5 +10,5 @@ data class ReconcileEntriesCommand(
     val accountId: AccountId? = null,
     val from: Instant,
     val to: Instant,
-    val tolerancePercent: Double? = null,
+    val tolerancePercent: BigDecimal? = null,
 )
